Separate fs management in its own file/class
